Move CodepointView class to its own header
