+++ src/openrct2/drawing/ScrollingText.cpp
@@ -10,6 +10,7 @@
 #include "ScrollingText.h"
 
 #include "../config/Config.h"
+#include "../core/CodepointView.hpp"
 #include "../core/String.hpp"
 #include "../interface/Colour.h"
 #include "../localisation/Formatter.h"
