+++ src/openrct2/paint/tile_element/Paint.LargeScenery.cpp
@@ -12,8 +12,8 @@
 #include "../../Game.h"
 #include "../../GameState.h"
 #include "../../config/Config.h"
+#include "../../core/CodepointView.hpp"
 #include "../../core/Numerics.hpp"
-#include "../../core/String.hpp"
 #include "../../core/UTF8.h"
 #include "../../interface/Viewport.h"
 #include "../../localisation/Formatting.h"
